+++ odbc/odbc_api.h
@@ -72,6 +72,12 @@ SQLRETURN SQLProcedureColumns(
     SQLSMALLINT name_length2, SQLCHAR* procedure_name,
     SQLSMALLINT name_length3, SQLCHAR* column_name,
     SQLSMALLINT name_length4);
+SQLRETURN SQLSpecialColumns(
+    SQLHSTMT statement_handle, SQLUSMALLINT identifier_type,
+    SQLCHAR* catalog_name, SQLSMALLINT name_length1,
+    SQLCHAR* schema_name, SQLSMALLINT name_length2,
+    SQLCHAR* table_name, SQLSMALLINT name_length3,
+    SQLUSMALLINT scope, SQLUSMALLINT nullable);
 SQLRETURN SQLTables(
     SQLHSTMT statement_handle, SQLCHAR* catalog_name,
     SQLSMALLINT name_length1, SQLCHAR* schema_name,